Refactor Comment Handling
Update the antlr grammar so that comments are parsed with their
surrounding whitespace, allowing us to more easily preserve the
comment formatting.

Fix lots of comment formatting edge cases.

Fix parsing of class body in which it wasn't visited properly, resulting
in comments outside the body to be moved inside the body.